+++ tests/Test/Cores/Ethernet/Mac/MacDepacketizer.hs
@@ -2,7 +2,7 @@
 {-# language NumericUnderscores #-}
 {-# language RecordWildCards #-}
 
-module Test.Cores.Ethernet.PreambleInserter where
+module Test.Cores.Ethernet.Mac.MacDepacketizer where
 
 -- base
 import Prelude
@@ -27,34 +27,31 @@ import Protocols.Hedgehog
 
 -- Me
 import Clash.Cores.Ethernet.Mac.EthernetTypes
-import Clash.Cores.Ethernet.Mac.Preamble ( preambleInserterC )
+import Clash.Cores.Ethernet.Mac.MacPacketizers ( macDepacketizerC )
 import Protocols.Extra.PacketStream
 
-import Test.Cores.Ethernet.Packetizer ( packetizerModel )
 import Test.Cores.Ethernet.Util
-
-import Data.List qualified as L
-
+import Test.Protocols.Extra.PacketStream.Depacketizer ( depacketizerModel )
 
 genVec :: (C.KnownNat n, 1 <= n) => Gen a -> Gen (C.Vec n a)
 genVec gen = sequence (C.repeat gen)
 
-preambleInserterPropertyGenerator
-  :: forall (dataWidth :: Nat) .
-     1 <= dataWidth
+macDepacketizerPropertyGenerator
+  :: forall (dataWidth :: Nat)
+   . 1 <= dataWidth
   => SNat dataWidth
   -> Property
-preambleInserterPropertyGenerator SNat =
+macDepacketizerPropertyGenerator SNat =
   propWithModelSingleDomain
     @C.System
     defExpectOptions
     (fmap fullPackets (Gen.list (Range.linear 1 100) genPackets))
     (C.exposeClockResetEnable model)
-    (C.exposeClockResetEnable @C.System preambleInserterC)
+    (C.exposeClockResetEnable @C.System macDepacketizerC)
     (===)
     where
-      model :: [PacketStreamM2S dataWidth ()] -> [PacketStreamM2S dataWidth ()]
-      model = packetizerModel (const ()) id . L.map (\x -> x { _meta = preamble })
+      model :: [PacketStreamM2S dataWidth ()] -> [PacketStreamM2S dataWidth EthernetHeader]
+      model = depacketizerModel const
       genPackets =
           PacketStreamM2S <$>
           genVec Gen.enumBounded <*>
@@ -63,28 +60,28 @@ preambleInserterPropertyGenerator SNat =
           Gen.enumBounded
 
 -- | n mod dataWidth ~ 1
-prop_preamble_inserter_d1 :: Property
-prop_preamble_inserter_d1 = preambleInserterPropertyGenerator d1
+prop_mac_depacketizer_d1 :: Property
+prop_mac_depacketizer_d1 = macDepacketizerPropertyGenerator d1
 
 -- | n mod dataWidth ~ 3
-prop_preamble_inserter_d3 :: Property
-prop_preamble_inserter_d3 = preambleInserterPropertyGenerator d3
+prop_mac_depacketizer_d3 :: Property
+prop_mac_depacketizer_d3 = macDepacketizerPropertyGenerator d3
 
 -- | n mod dataWidth ~ 0
-prop_preamble_inserter_d7 :: Property
-prop_preamble_inserter_d7 = preambleInserterPropertyGenerator d7
+prop_mac_depacketizer_d7 :: Property
+prop_mac_depacketizer_d7 = macDepacketizerPropertyGenerator d7
 
 -- | dataWidth < header byte size
-prop_preamble_inserter_d9 :: Property
-prop_preamble_inserter_d9 = preambleInserterPropertyGenerator d9
+prop_mac_depacketizer_d9 :: Property
+prop_mac_depacketizer_d9 = macDepacketizerPropertyGenerator d9
 
 -- | dataWidth ~ header byte size
-prop_preamble_inserter_d14 :: Property
-prop_preamble_inserter_d14 = preambleInserterPropertyGenerator d14
+prop_mac_depacketizer_d14 :: Property
+prop_mac_depacketizer_d14 = macDepacketizerPropertyGenerator d14
 
 -- | dataWidth > header byte size
-prop_preamble_inserter_d15 :: Property
-prop_preamble_inserter_d15 = preambleInserterPropertyGenerator d15
+prop_mac_depacketizer_d15 :: Property
+prop_mac_depacketizer_d15 = macDepacketizerPropertyGenerator d15
 
 tests :: TestTree
 tests =
