+++ tests/Test/Cores/Ethernet/Mac/PreambleInserter.hs
@@ -2,7 +2,7 @@
 {-# language NumericUnderscores #-}
 {-# language RecordWildCards #-}
 
-module Test.Cores.Ethernet.PreambleStripper where
+module Test.Cores.Ethernet.Mac.PreambleInserter where
 
 -- base
 import Prelude
@@ -27,11 +27,11 @@ import Protocols.Hedgehog
 
 -- Me
 import Clash.Cores.Ethernet.Mac.EthernetTypes
-import Clash.Cores.Ethernet.Mac.Preamble ( preambleStripperC )
+import Clash.Cores.Ethernet.Mac.Preamble ( preambleInserterC )
 import Protocols.Extra.PacketStream
 
-import Test.Cores.Ethernet.Depacketizer ( depacketizerModel )
 import Test.Cores.Ethernet.Util
+import Test.Protocols.Extra.PacketStream.Packetizer ( packetizerModel )
 
 import Data.List qualified as L
 
@@ -39,29 +39,22 @@ import Data.List qualified as L
 genVec :: (C.KnownNat n, 1 <= n) => Gen a -> Gen (C.Vec n a)
 genVec gen = sequence (C.repeat gen)
 
-preambleStripperPropertyGenerator
+preambleInserterPropertyGenerator
   :: forall (dataWidth :: Nat) .
      1 <= dataWidth
   => SNat dataWidth
   -> Property
-preambleStripperPropertyGenerator SNat =
+preambleInserterPropertyGenerator SNat =
   propWithModelSingleDomain
     @C.System
     defExpectOptions
     (fmap fullPackets (Gen.list (Range.linear 1 100) genPackets))
     (C.exposeClockResetEnable model)
-    (C.exposeClockResetEnable @C.System preambleStripperC)
+    (C.exposeClockResetEnable @C.System preambleInserterC)
     (===)
     where
-      model ps = validateAll (depacketizerModel const ps)
-
-      validateAll :: [PacketStreamM2S dataWidth Preamble] -> [PacketStreamM2S dataWidth ()]
-      validateAll ps = L.concatMap validatePreamble (chunkByPacket ps)
-
-      validatePreamble :: [PacketStreamM2S dataWidth Preamble] -> [PacketStreamM2S dataWidth ()]
-      validatePreamble ps = if C.last (_meta $ Prelude.head ps) == startFrameDelimiter
-                            then L.map (\p -> p {_meta = ()}) ps
-                            else []
+      model :: [PacketStreamM2S dataWidth ()] -> [PacketStreamM2S dataWidth ()]
+      model = packetizerModel (const ()) id . L.map (\x -> x { _meta = preamble })
       genPackets =
           PacketStreamM2S <$>
           genVec Gen.enumBounded <*>
@@ -70,28 +63,28 @@ preambleStripperPropertyGenerator SNat =
           Gen.enumBounded
 
 -- | n mod dataWidth ~ 1
-prop_preamble_stripper_d1 :: Property
-prop_preamble_stripper_d1 = preambleStripperPropertyGenerator d1
+prop_preamble_inserter_d1 :: Property
+prop_preamble_inserter_d1 = preambleInserterPropertyGenerator d1
 
 -- | n mod dataWidth ~ 3
-prop_preamble_stripper_d5 :: Property
-prop_preamble_stripper_d5 = preambleStripperPropertyGenerator d5
+prop_preamble_inserter_d3 :: Property
+prop_preamble_inserter_d3 = preambleInserterPropertyGenerator d3
 
 -- | n mod dataWidth ~ 0
-prop_preamble_stripper_d4 :: Property
-prop_preamble_stripper_d4 = preambleStripperPropertyGenerator d4
+prop_preamble_inserter_d7 :: Property
+prop_preamble_inserter_d7 = preambleInserterPropertyGenerator d7
 
 -- | dataWidth < header byte size
-prop_preamble_stripper_d7 :: Property
-prop_preamble_stripper_d7 = preambleStripperPropertyGenerator d7
+prop_preamble_inserter_d9 :: Property
+prop_preamble_inserter_d9 = preambleInserterPropertyGenerator d9
 
 -- | dataWidth ~ header byte size
-prop_preamble_stripper_d8 :: Property
-prop_preamble_stripper_d8 = preambleStripperPropertyGenerator d8
+prop_preamble_inserter_d14 :: Property
+prop_preamble_inserter_d14 = preambleInserterPropertyGenerator d14
 
 -- | dataWidth > header byte size
-prop_preamble_stripper_d9 :: Property
-prop_preamble_stripper_d9 = preambleStripperPropertyGenerator d9
+prop_preamble_inserter_d15 :: Property
+prop_preamble_inserter_d15 = preambleInserterPropertyGenerator d15
 
 tests :: TestTree
 tests =
