+++ tests/unittests.hs
@@ -2,54 +2,54 @@ import Prelude
 
 import Test.Tasty
 
-import Test.Cores.Arp.ArpManager qualified
-import Test.Cores.Arp.ArpReceiver qualified
-import Test.Cores.Arp.ArpTable qualified
-import Test.Cores.Arp.ArpTransmitter qualified
-import Test.Cores.Ethernet.AsyncFIFO qualified
-import Test.Cores.Ethernet.DownConverter qualified
-import Test.Cores.Ethernet.EthernetTypes qualified
-import Test.Cores.Ethernet.FrameCheckSequence qualified
-import Test.Cores.Ethernet.InternetChecksum qualified
-import Test.Cores.Ethernet.InterpacketGapInserter qualified
-import Test.Cores.Ethernet.IpPacketizer qualified
-import Test.Cores.Ethernet.MacDepacketizer qualified
-import Test.Cores.Ethernet.MacPacketizer qualified
-import Test.Cores.Ethernet.PacketArbiter qualified
-import Test.Cores.Ethernet.PacketBuffer qualified
-import Test.Cores.Ethernet.PacketDispatcher qualified
-import Test.Cores.Ethernet.PacketStream qualified
-import Test.Cores.Ethernet.PaddingInserter qualified
-import Test.Cores.Ethernet.PreambleInserter qualified
-import Test.Cores.Ethernet.PreambleStripper qualified
-import Test.Cores.Ethernet.UpConverter qualified
-import Test.Cores.IP.Icmp qualified
-import Test.Cores.IP.IPDepacketizer qualified
+import Test.Cores.Ethernet.Arp.ArpManager qualified
+import Test.Cores.Ethernet.Arp.ArpReceiver qualified
+import Test.Cores.Ethernet.Arp.ArpTable qualified
+import Test.Cores.Ethernet.Arp.ArpTransmitter qualified
+import Test.Cores.Ethernet.Icmp qualified
+import Test.Cores.Ethernet.IP.InternetChecksum qualified
+import Test.Cores.Ethernet.IP.IPDepacketizer qualified
+import Test.Cores.Ethernet.IP.IPPacketizer qualified
+import Test.Cores.Ethernet.Mac.AsyncFIFO qualified
+import Test.Cores.Ethernet.Mac.EthernetTypes qualified
+import Test.Cores.Ethernet.Mac.FrameCheckSequence qualified
+import Test.Cores.Ethernet.Mac.InterpacketGapInserter qualified
+import Test.Cores.Ethernet.Mac.MacDepacketizer qualified
+import Test.Cores.Ethernet.Mac.MacPacketizer qualified
+import Test.Cores.Ethernet.Mac.PaddingInserter qualified
+import Test.Cores.Ethernet.Mac.PreambleInserter qualified
+import Test.Cores.Ethernet.Mac.PreambleStripper qualified
+import Test.Protocols.Extra.PacketStream.UpConverter qualified
+import Test.Protocols.Extra.PacketStream qualified
+import Test.Protocols.Extra.PacketStream.DownConverter qualified
+import Test.Protocols.Extra.PacketStream.PacketArbiter qualified
+import Test.Protocols.Extra.PacketStream.PacketBuffer qualified
+import Test.Protocols.Extra.PacketStream.PacketDispatcher qualified
 
 main :: IO ()
 main = defaultMain $ testGroup "."
-  [ Test.Cores.Arp.ArpManager.tests
-    , Test.Cores.Arp.ArpReceiver.tests
-    , Test.Cores.Arp.ArpTable.tests
-    , Test.Cores.Arp.ArpTransmitter.tests
-    , Test.Cores.Ethernet.AsyncFIFO.tests
-    , Test.Cores.Ethernet.PacketArbiter.tests
-    , Test.Cores.Ethernet.PacketStream.tests
-    , Test.Cores.Ethernet.PaddingInserter.tests
-    , Test.Cores.Ethernet.UpConverter.tests
-    , Test.Cores.Ethernet.DownConverter.tests
-    , Test.Cores.Ethernet.PacketBuffer.tests
-    , Test.Cores.Ethernet.PacketDispatcher.tests
-    , Test.Cores.Ethernet.PreambleInserter.tests
-    , Test.Cores.Ethernet.PreambleStripper.tests
-    , Test.Cores.Ethernet.InterpacketGapInserter.tests
-    , Test.Cores.Ethernet.MacDepacketizer.tests
-    , Test.Cores.Ethernet.MacPacketizer.tests
-    , Test.Cores.Ethernet.InternetChecksum.tests
-    , Test.Cores.Ethernet.FrameCheckSequence.testsValidate
-    , Test.Cores.Ethernet.FrameCheckSequence.testsInsert
-    , Test.Cores.Ethernet.EthernetTypes.tests
-    , Test.Cores.IP.Icmp.tests
-    , Test.Cores.IP.IPDepacketizer.tests
-    , Test.Cores.Ethernet.IpPacketizer.tests
+  [ Test.Cores.Ethernet.Arp.ArpManager.tests
+  , Test.Cores.Ethernet.Arp.ArpReceiver.tests
+  , Test.Cores.Ethernet.Arp.ArpTable.tests
+  , Test.Cores.Ethernet.Arp.ArpTransmitter.tests
+  , Test.Cores.Ethernet.Mac.AsyncFIFO.tests
+  , Test.Protocols.Extra.PacketStream.PacketArbiter.tests
+  , Test.Protocols.Extra.PacketStream.tests
+  , Test.Cores.Ethernet.Mac.PaddingInserter.tests
+  , Test.Protocols.Extra.PacketStream.UpConverter.tests
+  , Test.Protocols.Extra.PacketStream.DownConverter.tests
+  , Test.Protocols.Extra.PacketStream.PacketBuffer.tests
+  , Test.Protocols.Extra.PacketStream.PacketDispatcher.tests
+  , Test.Cores.Ethernet.Mac.PreambleInserter.tests
+  , Test.Cores.Ethernet.Mac.PreambleStripper.tests
+  , Test.Cores.Ethernet.Mac.InterpacketGapInserter.tests
+  , Test.Cores.Ethernet.Mac.MacDepacketizer.tests
+  , Test.Cores.Ethernet.Mac.MacPacketizer.tests
+  , Test.Cores.Ethernet.IP.InternetChecksum.tests
+  , Test.Cores.Ethernet.Mac.FrameCheckSequence.testsValidate
+  , Test.Cores.Ethernet.Mac.FrameCheckSequence.testsInsert
+  , Test.Cores.Ethernet.Mac.EthernetTypes.tests
+  , Test.Cores.Ethernet.Icmp.tests
+  , Test.Cores.Ethernet.IP.IPDepacketizer.tests
+  , Test.Cores.Ethernet.IP.IPPacketizer.tests
   ]
